+++ kiwi.js
@@ -1129,7 +1129,12 @@ var BLOCK_PARSE_RE = /^block\s+([^\s]+)(?:\s+(append|prepend))?$/;
 
 module.exports.tags = {};
 var blockTag = module.exports.tags.block = {};
+var parentTag = module.exports.tags.parent = {};
 
+
+/**
+ * Block tag
+ */
 
 /**
  * Basic tag settings
@@ -1162,14 +1167,28 @@ blockTag.compile = function(token, compiledContents,
 
   var name = parsed[1];
   var mode = parsed[2];
+  
+  //console.log(compiledContents);
 
   var compiled =   '(function(parentAcc) {'
+                 +   'var __acc = [];'
+                 +   compiledContents
+                 +   'var __currentBlock = __acc.join("");'
+                 +   'if(!_.isUndefined(__blocks["' + name + '"])) {'
+                 +     '__tmp = new String(__blocks["' + name + '"].replace('
+                 +       '"{{parent}}", __currentBlock'
+                 +     '));'
+                 +     'if(__blocks["' + name + '"].mode) {'
+                 +       '__tmp.mode = __blocks["' + name + '"].mode;'
+                 +     '}'
+                 +     '__blocks["' + name + '"] = __tmp;'
+                 +   '}'
                  +   'var __acc = [];'
                  +   'if(_.isUndefined(__blocks["' + name + '"]) || __blocks["' + name + '"].mode) {'
                  +     'if(__blocks["' + name + '"] && __blocks["' + name + '"].mode == "append") {'
                  +       '__acc.push(__blocks["' + name + '"]);'
                  +     '}'
-                 +     compiledContents
+                 +     '__acc.push(__currentBlock);'
                  +     'if(__blocks["' + name + '"] && __blocks["' + name + '"].mode == "prepend") {'
                  +       '__acc.push(__blocks["' + name + '"]);'
                  +     '}'
@@ -1183,6 +1202,31 @@ blockTag.compile = function(token, compiledContents,
                  + '})(__acc);'
 
   callback(null, compiled);
+}
+
+
+/**
+ * Parent tag
+ */
+ 
+/**
+ * Compile `token` with `compiledContents` to JavaScript, and invoke
+ * `callback(err, compiled)`.
+ *
+ * @param {BlockToken} token
+ * @param {String} compiledContents
+ * @param {Function} callback
+ * @api private
+ */
+
+parentTag.compile = function(token, compiledContents, compiler, callback) {
+  if(token.parent.tagType.tagName !== 'block') {
+    return callback(new Error('Compilation error: `parent` tag should be immediate child of a `block` tag.'));
+  }
+  
+  token.parent.hasParentTag = true;
+
+  callback(null, '__acc.push("{{parent}}");');
 }
 
 }); // module: tags/block.js