+++ test/cases/tag.block.js
@@ -13,7 +13,7 @@ describe('Block tag', function() {
   var template;
 
   beforeEach(function() {
-    template = new Template({cache: true});
+    template = new Template({cache: false});
   });
 
   it('should properky handle `append` and `prepend` directives', function(done) {
@@ -30,5 +30,21 @@ describe('Block tag', function() {
     }
 
     template.loadFile(__dirname + '/../fixtures/block.kiwi', onLoaded);
+  });
+  
+  it('should properky handle `parent` tag', function(done) {
+
+    function onLoaded(err, data) {
+      if(err) return done(err);
+      template.render({}, onRendered)
+    }
+
+    function onRendered(err, rendered) {
+      if(err) return done(err);
+      rendered.trim().should.equal('foo bar woo doo loo roo');
+      done();
+    }
+
+    template.loadFile(__dirname + '/../fixtures/parent.kiwi', onLoaded);
   });
 });